move remaining student-based clis into gob-cli
