--- modules/gob-examine-student-cli/.eslintrc.yml
@@ -1,2 +1,0 @@
----
-extends: ../gob-cli/.eslintrc.yml
