--- modules/gob-examine-student-cli/evaluate
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-
-require('./../../cli/lib/init')
-require('./evaluate.js').cli()
