--- modules/gob-object-student-cli/olaf-count-courses-with-ges
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-
-require('./lib/init')
-require('./olaf-count-courses-with-ges.js').cli()
