--- modules/gob-object-student-cli/olaf-find-student-courses
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-
-require('./../../cli/lib/init')
-require('./olaf-find-student-courses.js').cli()
