--- modules/gob-object-student-cli/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@gob/object-student-cli",
-  "version": "3.0.0-rc.2",
-  "main": "index.js",
-  "repository": "hawkrives/gobbldygook",
-  "author": "Hawken MacKay Rives",
-  "license": "MIT",
-  "dependencies": {
-    "@gob/cli": "^3.0.0-rc.2",
-    "@gob/examine-student": "^3.0.0-rc.2",
-    "chalk": "^2.4.1",
-    "graceful-fs": "^4.1.11",
-    "js-yaml": "^3.12.0",
-    "junk": "^2.1.0",
-    "lodash": "^4.17.11",
-    "nomnom": "^1.8.1",
-    "pify": "^3.0.0",
-    "plur": "^3.0.1",
-    "sparkly": "^3.1.2"
-  }
-}
