--- modules/gob-object-student-cli/students-benchmark
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-
-require('./lib/init')
-require('./students-benchmark.js').cli()
