--- modules/gob-object-student-cli/test-student
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-
-require('./../../cli/lib/init')
-require('./test-student.js').cli()
